tests/appmanifest/test.html

<!DOCTYPE html>
<!--
Copyright (c) 2015 Intel Corporation.

Redistribution and use in source and binary forms, with or without modification,
are permitted provided that the following conditions are met:

* Redistributions of works must retain the original copyright notice, this list
  of conditions and the following disclaimer.
* Redistributions in binary form must reproduce the original copyright notice,
  this list of conditions and the following disclaimer in the documentation
  and/or other materials provided with the distribution.
* Neither the name of Intel Corporation nor the names of its contributors
  may be used to endorse or promote products derived from this work without
  specific prior written permission.

THIS SOFTWARE IS PROVIDED BY INTEL CORPORATION "AS IS"
AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
ARE DISCLAIMED. IN NO EVENT SHALL INTEL CORPORATION BE LIABLE FOR ANY DIRECT,
INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY
OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
-->

<title>Clients.get()</title>
<link rel="author" title="Intel" href="http://www.intel.com">
<link rel="author" title="Xiuqi Jiang" href="mailto:xiuqix.jiang@intel.com">
<link rel="help" href="https://slightlyoff.github.io/ServiceWorker/spec/service_worker/index.html#clients-get-method">
<script src="resources/testharness.js"></script>
<script src="resources/testharnessreport.js"></script>
<script src="resources/test-helpers.js"></script>
<div id="log"></div>
<script>
var scope = 'resources/clients-get-frame.html';
var t = async_test('Test Clients.get()');
var clientIds = [];
var frame;
t.step(function() {
    service_worker_unregister_and_register(
        t, 'resources/clients-get-worker.js', scope)
      .then(function(registration) {
          return wait_for_state(t, registration.installing, 'activated');
        })
      .then(function() {
          return with_iframe(scope + '#1');
        })
      .then(function(frame1) {
          frame1.focus();
          return wait_for_clientId();
        })
      .then(function(clientId) {
          clientIds.push(clientId);
          return with_iframe(scope + '#2');
        })
      .then(function(frame2) {
          frame = frame2;
          return wait_for_clientId();
        })
      .then(function(clientId) {
          clientIds.push(clientId);
          var channel = new MessageChannel();
          channel.port1.onmessage = t.step_func(on_message);
          frame.contentWindow.navigator.serviceWorker.controller.postMessage(
              {port:channel.port2, clientIds:clientIds,
               message: 'get_client_ids'}, [channel.port2]);
        })
      .catch(unreached_rejection(t));
  });

function wait_for_clientId() {
  return new Promise(function(resolve, reject) {
      function get_client_id(e) {
        window.removeEventListener("message", get_client_id);
        resolve(e.data.clientId);
      }
      window.addEventListener("message", get_client_id, false);
    });
}

var expected = [
    /* visibilityState, focused, url, frameType */
    ['visible', true, new URL(scope + '#1', location).toString(), 'nested'],
    ['visible', false, new URL(scope + '#2', location).toString(), 'nested'],
    undefined
];

function on_message(e) {
  assert_equals(e.data.length, 3);
  assert_array_equals(e.data[0], expected[0]);
  assert_array_equals(e.data[1], expected[1]);
  assert_equals(e.data[2], expected[2]);
  service_worker_unregister_and_done(t, scope);
}
</script>
